Update SPLA_Code.ino
Added a Buffer for the FFT, which allows for a FFT_SIZE bigger than 1024.
In the previous code, the FFT_SIZE is dependent on the i2s bufferlength, which can at most be 1024.
Made the appropriate changes in the main loop to accommodate this. The new FFT_SIZE is 8192 because it allows for all the frequency bands to be represented (including 25Hz). If 25Hz is not needed, the FFT_SIZE can be halved.